CD-8904 | PR changes
